+++ src/snowcli/cli/streamlit/commands.py
@@ -1,4 +1,5 @@
 import logging
+
 import typer
 from pathlib import Path
 from typing import Optional
@@ -7,15 +8,10 @@ from snowcli.cli.common.decorators import global_options_with_connection, global
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
 from snowcli.cli.streamlit.manager import StreamlitManager
 from snowcli.output.decorators import with_output
-from snowcli.cli.snowpark_shared import (
-    CheckAnacondaForPyPiDependencies,
-    PackageNativeLibrariesOption,
-    PyPiDownloadOption,
-)
+
 from snowcli.output.types import (
     CommandResult,
     QueryResult,
-    CollectionResult,
     SingleQueryResult,
     MessageResult,
     MultipleResults,
@@ -28,6 +24,13 @@ app = typer.Typer(
     help="Manages Streamlit in Snowflake.",
 )
 log = logging.getLogger(__name__)
+
+
+StageNameOption: str = typer.Option(
+    "streamlit",
+    "--stage",
+    help="Stage name where are Streamlit files.",
+)
 
 
 @app.command("init")
@@ -76,45 +79,6 @@ def streamlit_describe(
     return result
 
 
-@app.command("create")
-@with_output
-@global_options_with_connection
-def streamlit_create(
-    name: str = typer.Argument(..., help="Name of streamlit to create."),
-    file: Path = typer.Option(
-        "streamlit_app.py",
-        file_okay=True,
-        help="Path on stage to the main file of Streamlit application.",
-    ),
-    from_stage: Optional[str] = typer.Option(
-        None,
-        help="Stage name from which to copy a Streamlit file.",
-    ),
-    use_packaging_workaround: bool = typer.Option(
-        False,
-        help="Whether to package all code and dependencies into a zip file. Valid values: `true`, `false` (default). You should use this only for a temporary workaround until native support is available.",
-    ),
-    **options,
-) -> CommandResult:
-    """
-    Creates a new Streamlit application object in Snowflake. The streamlit is created in database and schema configured in the connection.
-    """
-    sm = StreamlitManager()
-    results = MultipleResults()
-    create_result = sm.create(
-        streamlit_name=name,
-        file=file,
-        stage=from_stage,
-        use_packaging_workaround=use_packaging_workaround,
-    )
-    results.add(SingleQueryResult(create_result))
-
-    checkout_result = sm.checkout(streamlit_name=name)
-    results.add(SingleQueryResult(checkout_result))
-
-    return results
-
-
 @app.command("share")
 @with_output
 @global_options_with_connection
@@ -150,7 +114,7 @@ def streamlit_drop(
 @with_output
 @global_options_with_connection
 def streamlit_deploy(
-    name: str = typer.Argument(..., help="Name of streamlit to deploy."),
+    streamlit_name: str = typer.Argument(..., help="Name of Streamlit to deploy."),
     file: Path = typer.Option(
         "streamlit_app.py",
         exists=True,
@@ -158,43 +122,53 @@ def streamlit_deploy(
         file_okay=True,
         help="Path of the Streamlit app file.",
     ),
+    stage: Optional[str] = StageNameOption,
+    environment_file: Path = typer.Option(
+        "environment.yml",
+        "--env-file",
+        help="Environment file to use.",
+        file_okay=True,
+        dir_okay=False,
+    ),
+    pages_dir: Path = typer.Option(
+        "pages",
+        "--pages-dir",
+        help="Directory with Streamlit pages",
+        file_okay=False,
+        dir_okay=True,
+    ),
+    query_warehouse: Optional[str] = typer.Option(
+        None, "--query-warehouse", help="Query warehouse for this Streamlit."
+    ),
+    replace: Optional[bool] = typer.Option(
+        False,
+        "--replace",
+        help="Replace the Streamlit if it already exists.",
+        is_flag=True,
+    ),
     open_: bool = typer.Option(
-        False,
-        "--open",
-        "-o",
-        help="Whether to open Streamlit in a browser. Valid values: `true`, `false`. Default: `false`.",
-    ),
-    use_packaging_workaround: bool = typer.Option(
-        False,
-        help="Whether to package all code and dependencies into a zip file. Valid values: `true`, `false` (default). You should use this only for a temporary workaround until native support is available.",
-    ),
-    packaging_workaround_includes_content: bool = typer.Option(
-        False,
-        help="Whether to package all code and dependencies into a zip file. Valid values: `true`, `false`. Default: `false`.",
-    ),
-    pypi_download: str = PyPiDownloadOption,
-    check_anaconda_for_pypi_deps: bool = CheckAnacondaForPyPiDependencies,
-    package_native_libraries: str = PackageNativeLibrariesOption,
-    excluded_anaconda_deps: str = typer.Option(
-        None,
-        help="List of comma-separated package names from `environment.yml` to exclude in the deployed app, particularly when Streamlit fails to import an Anaconda package at runtime. Be aware that excluding files might the risk of runtime errors).",
+        False, "--open", help="Whether to open Streamlit in a browser.", is_flag=True
     ),
     **options,
 ) -> CommandResult:
     """
-    Creates a Streamlit app package for deployment.
+    Uploads local files to specified stage and creates a Streamlit dashboard using the files. You must specify the
+    main python file. By default, the command will upload environment.yml and pages/ folder  if present. If you
+    don't provide any stage name then 'streamlit' stage will be used. If provided stage will be created if it does
+    not exist.
+    You can modify the behaviour using flags. For details check help information.
     """
-    result = StreamlitManager().deploy(
-        streamlit_name=name,
-        file=file,
-        open_in_browser=open_,
-        use_packaging_workaround=use_packaging_workaround,
-        packaging_workaround_includes_content=packaging_workaround_includes_content,
-        pypi_download=pypi_download,
-        check_anaconda_for_pypi_deps=check_anaconda_for_pypi_deps,
-        package_native_libraries=package_native_libraries,
-        excluded_anaconda_deps=excluded_anaconda_deps,
+    url = StreamlitManager().deploy(
+        streamlit_name=streamlit_name,
+        environment_file=environment_file,
+        pages_dir=pages_dir,
+        stage_name=stage,
+        main_file=file,
+        replace=replace,
+        warehouse=query_warehouse,
     )
-    if result is not None:
-        return MessageResult(result)
-    return MessageResult("Done")
+
+    if open_:
+        typer.launch(url)
+
+    return MessageResult(f"Streamlit successfully deployed and available under {url}")