+++ tests/streamlit/test_commands.py
@@ -11,101 +11,6 @@ STREAMLIT_NAME = "test_streamlit"
 
 
 @mock.patch("snowflake.connector.connect")
-def test_create_streamlit(mock_connector, runner, mock_ctx):
-    ctx = mock_ctx()
-    mock_connector.return_value = ctx
-
-    with NamedTemporaryFile(suffix=".py") as file:
-        result = runner.invoke_with_config(
-            ["streamlit", "create", STREAMLIT_NAME, "--file", file.name]
-        )
-
-        assert result.exit_code == 0, result.output
-        assert ctx.get_query() == dedent(
-            f"""
-                create streamlit {STREAMLIT_NAME}
-            
-                MAIN_FILE = '{os.path.basename(file.name)}'
-                QUERY_WAREHOUSE = MockWarehouse;
-                
-                alter streamlit {STREAMLIT_NAME} checkout;
-            """
-        )
-
-
-@mock.patch("snowflake.connector.connect")
-def test_create_streamlit_with_use_packaging_workaround(
-    mock_connector, runner, mock_ctx
-):
-    ctx = mock_ctx()
-    mock_connector.return_value = ctx
-
-    with NamedTemporaryFile(suffix=".py") as file:
-        result = runner.invoke_with_config(
-            [
-                "streamlit",
-                "create",
-                STREAMLIT_NAME,
-                "--file",
-                file.name,
-                "--use-packaging-workaround",
-            ]
-        )
-
-        assert result.exit_code == 0, result.output
-        assert ctx.get_query() == dedent(
-            f"""
-                create streamlit {STREAMLIT_NAME}
-            
-                MAIN_FILE = 'streamlit_app_launcher.py'
-                QUERY_WAREHOUSE = MockWarehouse;
-                
-                alter streamlit {STREAMLIT_NAME} checkout;
-            """
-        )
-
-
-@pytest.mark.parametrize(
-    "stage, expected",
-    [
-        ("stage_name", "@stage_name"),
-        ("snow://stage_dots", "snow://stage_dots"),
-    ],
-)
-@mock.patch("snowflake.connector.connect")
-def test_create_streamlit_with_from_stage(
-    mock_connector, runner, mock_ctx, stage, expected
-):
-    ctx = mock_ctx()
-    mock_connector.return_value = ctx
-
-    with NamedTemporaryFile(suffix=".py") as file:
-        result = runner.invoke(
-            [
-                "streamlit",
-                "create",
-                STREAMLIT_NAME,
-                "--file",
-                file.name,
-                "--from-stage",
-                stage,
-            ]
-        )
-
-        assert result.exit_code == 0, result.output
-        assert ctx.get_query() == dedent(
-            f"""
-                create streamlit {STREAMLIT_NAME}
-                FROM {expected}
-                MAIN_FILE = '{os.path.basename(file.name)}'
-                QUERY_WAREHOUSE = MockWarehouse;
-                
-                alter streamlit {STREAMLIT_NAME} checkout;
-            """
-        )
-
-
-@mock.patch("snowflake.connector.connect")
 def test_list_streamlit(mock_connector, runner, mock_ctx):
     ctx = mock_ctx()
     mock_connector.return_value = ctx
@@ -130,17 +35,18 @@ def test_describe_streamlit(mock_connector, runner, mock_ctx):
     ]
 
 
-@mock.patch("snowcli.cli.streamlit.manager.typer")
-@mock.patch("snowcli.cli.streamlit.manager.StageManager")
-@mock.patch("snowflake.connector.connect")
-def test_deploy_streamlit(
-    mock_connector, mock_stage_manager, mock_typer, mock_cursor, runner, mock_ctx
-):
-    ctx = mock_ctx(
-        mock_cursor(
-            rows=["snowflake.com"], columns=["SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME"]
-        )
+def _put_query(source: str, dest: str):
+    return dedent(
+        f"put file://{source} {dest} auto_compress=false parallel=4 overwrite=True"
     )
+
+
+@mock.patch("snowcli.cli.streamlit.commands.typer")
+@mock.patch("snowflake.connector.connect")
+def test_deploy_streamlit_single_file(
+    mock_connector, mock_typer, mock_cursor, runner, mock_ctx
+):
+    ctx = mock_ctx()
     mock_connector.return_value = ctx
 
     with NamedTemporaryFile(suffix=".py") as file:
@@ -148,82 +54,159 @@ def test_deploy_streamlit(
             ["streamlit", "deploy", STREAMLIT_NAME, "--file", file.name, "--open"]
         )
 
-        assert result.exit_code == 0, result.output
-        assert (
-            ctx.get_query()
-            == f"call SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME('{STREAMLIT_NAME}')"
-        )
-        mock_stage_manager().put.assert_called_once_with(
-            file.name,
-            f"snow://streamlit/MockDatabase.MockSchema.{STREAMLIT_NAME}/default_checkout",
-            4,
-            True,
-        )
-        mock_typer.launch.assert_called_once_with(
-            f"https://account.test.region.aws.snowflakecomputing.com/test.region.aws/account/#/streamlit-apps/MOCKDATABASE.MOCKSCHEMA.{STREAMLIT_NAME.upper()}"
-        )
+    assert result.exit_code == 0, result.output
+    assert ctx.get_queries() == [
+        "create stage if not exists MOCKDATABASE.MOCKSCHEMA.STREAMLIT",
+        _put_query(file.name, "@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/test_streamlit"),
+        dedent(
+            f"""
+    CREATE  STREAMLIT {STREAMLIT_NAME}
+    ROOT_LOCATION = '@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/{STREAMLIT_NAME}'
+    MAIN_FILE = '{Path(file.name).name}'
 
+    """
+        ),
+        f"call SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME('{STREAMLIT_NAME}')",
+    ]
 
-@mock.patch("snowcli.cli.streamlit.manager.snowpark_package")
-@mock.patch("snowcli.cli.streamlit.manager.StageManager")
-@mock.patch("snowflake.connector.connect")
-def test_deploy_streamlit_with_packaging_workaround(
-    mock_connector,
-    mock_stage_manager,
-    mock_snowpark_package,
-    mock_cursor,
-    runner,
-    mock_ctx,
-    temp_dir,
-):
-    ctx = mock_ctx(
-        mock_cursor(
-            rows=["snowflake.com"], columns=["SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME"]
-        )
+    mock_typer.launch.assert_called_once_with(
+        f"https://account.test.region.aws.snowflakecomputing.com/test.region.aws/account/#/streamlit-apps/MOCKDATABASE.MOCKSCHEMA.{STREAMLIT_NAME.upper()}"
     )
+
+
+@mock.patch("snowflake.connector.connect")
+def test_deploy_streamlit_all_files_default_stage(
+    mock_connector, mock_cursor, runner, mock_ctx, project_file
+):
+    ctx = mock_ctx()
     mock_connector.return_value = ctx
 
-    with NamedTemporaryFile(suffix=".py") as file:
-        result = runner.invoke_with_config(
+    with project_file("example_streamlit") as pdir:
+        result = runner.invoke(
+            ["streamlit", "deploy", STREAMLIT_NAME, "--file", "main.py"]
+        )
+
+    root_path = f"@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/{STREAMLIT_NAME}"
+    assert result.exit_code == 0, result.output
+    assert ctx.get_queries() == [
+        "create stage if not exists MOCKDATABASE.MOCKSCHEMA.STREAMLIT",
+        _put_query("main.py", root_path),
+        _put_query("environment.yml", root_path),
+        _put_query("pages/*", f"{root_path}/pages"),
+        dedent(
+            f"""
+    CREATE  STREAMLIT {STREAMLIT_NAME}
+    ROOT_LOCATION = '@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/{STREAMLIT_NAME}'
+    MAIN_FILE = 'main.py'
+
+    """
+        ),
+        f"call SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME('{STREAMLIT_NAME}')",
+    ]
+
+
+@mock.patch("snowflake.connector.connect")
+def test_deploy_streamlit_all_files_users_stage(
+    mock_connector, mock_cursor, runner, mock_ctx, project_file
+):
+    ctx = mock_ctx()
+    mock_connector.return_value = ctx
+
+    with project_file("example_streamlit") as pdir:
+        result = runner.invoke(
             [
                 "streamlit",
                 "deploy",
                 STREAMLIT_NAME,
                 "--file",
-                file.name,
-                "--use-packaging-workaround",
+                "main.py",
+                "--stage",
+                "MY_FANCY_STAGE",
             ]
         )
 
-        assert result.exit_code == 0, result.output
-        assert (
-            ctx.get_query()
-            == f"call SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME('{STREAMLIT_NAME}')"
-        )
-        mock_snowpark_package.assert_called_once_with("ask", True, "ask")
+    root_path = f"@MOCKDATABASE.MOCKSCHEMA.MY_FANCY_STAGE/{STREAMLIT_NAME}"
+    assert result.exit_code == 0, result.output
+    assert ctx.get_queries() == [
+        "create stage if not exists MOCKDATABASE.MOCKSCHEMA.MY_FANCY_STAGE",
+        _put_query("main.py", root_path),
+        _put_query("environment.yml", root_path),
+        _put_query("pages/*", f"{root_path}/pages"),
+        dedent(
+            f"""
+    CREATE  STREAMLIT {STREAMLIT_NAME}
+    ROOT_LOCATION = '@MOCKDATABASE.MOCKSCHEMA.MY_FANCY_STAGE/{STREAMLIT_NAME}'
+    MAIN_FILE = 'main.py'
 
-        mock_stage_manager().put.assert_has_calls(
-            [
-                call(
-                    "app.zip",
-                    f"{STREAMLIT_NAME}_stage",
-                    4,
-                    True,
-                ),
-                call(
-                    "streamlit_app_launcher.py",
-                    f"{STREAMLIT_NAME}_stage",
-                    4,
-                    True,
-                ),
-                call(
-                    file.name,
-                    f"snow://streamlit/MockDatabase.MockSchema.{STREAMLIT_NAME}/default_checkout",
-                    4,
-                    True,
-                ),
-            ]
+    """
+        ),
+        f"call SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME('{STREAMLIT_NAME}')",
+    ]
+
+
+@mock.patch("snowflake.connector.connect")
+def test_deploy_streamlit_main_and_environment_files(
+    mock_connector, mock_cursor, runner, mock_ctx, project_file
+):
+    ctx = mock_ctx()
+    mock_connector.return_value = ctx
+
+    with project_file("example_streamlit") as pdir:
+
+        (pdir / "pages" / "my_page.py").unlink()
+        (pdir / "pages").rmdir()
+
+        result = runner.invoke(
+            ["streamlit", "deploy", STREAMLIT_NAME, "--file", "main.py"]
         )
+
+    root_path = f"@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/{STREAMLIT_NAME}"
+    assert result.exit_code == 0, result.output
+    assert ctx.get_queries() == [
+        "create stage if not exists MOCKDATABASE.MOCKSCHEMA.STREAMLIT",
+        _put_query("main.py", root_path),
+        _put_query("environment.yml", root_path),
+        dedent(
+            f"""
+    CREATE  STREAMLIT {STREAMLIT_NAME}
+    ROOT_LOCATION = '@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/{STREAMLIT_NAME}'
+    MAIN_FILE = 'main.py'
+
+    """
+        ),
+        f"call SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME('{STREAMLIT_NAME}')",
+    ]
+
+
+@mock.patch("snowflake.connector.connect")
+def test_deploy_streamlit_main_and_pages_files(
+    mock_connector, mock_cursor, runner, mock_ctx, project_file
+):
+    ctx = mock_ctx()
+    mock_connector.return_value = ctx
+
+    with project_file("example_streamlit") as pdir:
+        (pdir / "environment.yml").unlink()
+        result = runner.invoke(
+            ["streamlit", "deploy", STREAMLIT_NAME, "--file", "main.py"]
+        )
+
+    root_path = f"@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/{STREAMLIT_NAME}"
+    assert result.exit_code == 0, result.output
+    assert ctx.get_queries() == [
+        "create stage if not exists MOCKDATABASE.MOCKSCHEMA.STREAMLIT",
+        _put_query("main.py", root_path),
+        _put_query("pages/*", f"{root_path}/pages"),
+        dedent(
+            f"""
+    CREATE  STREAMLIT {STREAMLIT_NAME}
+    ROOT_LOCATION = '@MOCKDATABASE.MOCKSCHEMA.STREAMLIT/{STREAMLIT_NAME}'
+    MAIN_FILE = 'main.py'
+
+    """
+        ),
+        f"call SYSTEM$GENERATE_STREAMLIT_URL_FROM_NAME('{STREAMLIT_NAME}')",
+    ]
 
 
 @mock.patch("snowflake.connector.connect")