+++ RELEASE-NOTES.md
@@ -13,10 +13,12 @@
 * Update function or procedure will upload function/procedure code to new path on stage. Previous code will remain under old path on stage.
 * Snowpark command `compute-pool` and its alias `cp` were replaced by `pool` command.
 * `snow snowpark registry` was replaced with `snow registry` command.
+* Removed `snow streamlit create` command. Streamlit can be deployd using `snow streamlit deploy`
 
 ## New additions
 * `--temporary-connection` flag, that allows you to connect, without anything declared in config file
 * `snow streamlit init` command that creates a new streamlit project.
+* `snow streamlit deploy` support pages and environment.yml files.
 
 ## Fixes and improvements
 * Adjust streamlit commands to PuPr syntax